refactor(components): reorganize ImageZoomDialog import and move to visuals directory
