+++ components/visuals/index.ts
@@ -8,5 +8,4 @@ export { default as LoadingScreen } from "./LoadingScreen";
 export { default as LoadingSkeleton } from "./LoadingSkeleton";
 export { default as CentralizedLoading } from "./CentralizedLoading";
 export { Background } from "./motion-background";
-
-
+export { ImageZoomDialog } from "./image-zoom-dialog";
